test(e2e): restructure e2e tests and fix test reliability
Split e2e tests into eth-mainnet and evm-networks folders with separate
CI workflows. Fix HashRouter URL prefix (/#/) in all page objects. Add
Infura/Alchemy RPC injection via localStorage for reliable test execution.
Fix multiple test assertions: NFT selectors, token header text, trait
name casing, strict mode violations, and pagination timing.
