+++ .env.example
@@ -1,0 +1,7 @@
+# RPC API Keys for E2E Tests
+# Copy this file to .env and fill in your keys.
+# At least one key is recommended for reliable e2e tests.
+# If no keys are set, tests will use default public RPCs (slower, may be rate-limited).
+
+INFURA_API_KEY=
+ALCHEMY_API_KEY=
