+++ package.json
@@ -39,6 +39,8 @@
 		"test": "vitest",
 		"test:run": "vitest run",
 		"test:e2e": "playwright test",
+		"test:e2e:eth-mainnet": "playwright test e2e/tests/eth-mainnet/",
+		"test:e2e:evm-networks": "playwright test e2e/tests/evm-networks/",
 		"test:e2e:ui": "playwright test --ui",
 		"test:e2e:debug": "playwright test --debug",
 		"dev": "bash scripts/run-test-env.sh",
